Compile a WSDL file with multiple includes into a single file
The `wsdl2file` tool takes a WSDL, walks through all of its includes,
and compiles it down to a single file. It handles quirks with Microsoft
WCF such as it sometimes using backslashes instead of slashes, and also
comes with a `--zeep` flag that attempts to change the schema to make
it possible for Zeep to use it (which amounts to converting NIEM
abstract types into concrete types).

I wrote this after spending hours on the internet looking at other
peoples' solutions and finding that none of them could deal with the
insane complexity that is the ECF5 standard. When I get a few spare
cycles, I am going to clean this up (add a few unit tests, some more
documentation, etc) and release it as open source so that future people
that want to do modern programming against bloated government
webservices don't have to go through the same hell that I did.